+++ cmd/get_imports.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 
 	"github.com/Masterminds/cookoo"
-	v "github.com/Masterminds/go-vcs"
+	v "github.com/Masterminds/vcs"
 )
 
 func init() {